Extract interaction contract read model
